# hg-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hangman Game API",
        "description": "REST API for the Hangman Game — handles authentication, game sessions, and leaderboard.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/games": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "List all games for the authenticated user",
                "operationId": "93736f40a135111a310e4b25ad61f30b",
                "responses": {
                    "200": {
                        "description": "Array of game states (word revealed for finished games)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GameState"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Start a new game",
                "operationId": "b8b9557618eb0f7b3650bdd4ab3a6487",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "animal",
                                            "country",
                                            "food",
                                            "plant",
                                            "sport"
                                        ],
                                        "example": "animal"
                                    },
                                    "minLetters": {
                                        "type": "integer",
                                        "maximum": 20,
                                        "minimum": 1,
                                        "example": 4
                                    },
                                    "maxLetters": {
                                        "type": "integer",
                                        "maximum": 20,
                                        "minimum": 1,
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "animal",
                                            "country",
                                            "food",
                                            "plant",
                                            "sport"
                                        ],
                                        "example": "animal"
                                    },
                                    "minLetters": {
                                        "type": "integer",
                                        "maximum": 20,
                                        "minimum": 1,
                                        "example": 4
                                    },
                                    "maxLetters": {
                                        "type": "integer",
                                        "maximum": 20,
                                        "minimum": 1,
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New game created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameState"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/games/{game}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Get a single game",
                "operationId": "18467eed630a90558e542ccea443776a",
                "parameters": [
                    {
                        "name": "game",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 42
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game state",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameState"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Games"
                ],
                "summary": "Delete a game",
                "operationId": "28d8f1ca4b27acfd97823915d3648dab",
                "parameters": [
                    {
                        "name": "game",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 42
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Game deleted"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/games/{game}/guesses": {
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Submit a letter guess",
                "operationId": "6bdc32f0c4fdc74e0b8a789ad4f37977",
                "parameters": [
                    {
                        "name": "game",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 42
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "letter"
                                ],
                                "properties": {
                                    "letter": {
                                        "type": "string",
                                        "maxLength": 1,
                                        "minLength": 1,
                                        "example": "A"
                                    },
                                    "elapsed_seconds": {
                                        "description": "Front-end clock reading used for speed-bonus scoring",
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 30,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "letter"
                                ],
                                "properties": {
                                    "letter": {
                                        "type": "string",
                                        "maxLength": 1,
                                        "minLength": 1,
                                        "example": "A"
                                    },
                                    "elapsed_seconds": {
                                        "description": "Front-end clock reading used for speed-bonus scoring",
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 30,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated game state",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GameState"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Game already finished or letter already guessed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GameState": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 42
                    },
                    "masked_word": {
                        "type": "string",
                        "example": "_ A _ G _ A _"
                    },
                    "word_length": {
                        "type": "integer",
                        "example": 7
                    },
                    "guessed_letters": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "A",
                            "G"
                        ]
                    },
                    "wrong_guesses": {
                        "type": "integer",
                        "example": 2
                    },
                    "max_wrong_guesses": {
                        "type": "integer",
                        "example": 6
                    },
                    "remaining_attempts": {
                        "type": "integer",
                        "example": 4
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "in_progress",
                            "won",
                            "lost"
                        ],
                        "example": "in_progress"
                    },
                    "score": {
                        "type": "integer",
                        "example": 0
                    },
                    "elapsed_seconds": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "hint": {
                        "type": "string",
                        "example": "A large grey mammal",
                        "nullable": true
                    },
                    "category": {
                        "type": "string",
                        "example": "animal"
                    },
                    "word": {
                        "description": "Only present when the game is finished",
                        "type": "string",
                        "example": "ELEPHANT",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use the Bearer token returned by /auth/login or /auth/register.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}